this code allows third-parties to write backup plugins.  These plugins show up when the user chooses to backup their secrets
Some notes:

I did not do a progress bar/dialog while waiting for the secrets to be sent
from an OBA during a restore, but it makes sense for one to be shown, for now
the UI just hangs (bad user experience)

There is no upgrade code to convert current secrets to the new JSON format.
Eventually upgrade code needs to be written, for now this branch must be used
with a clean slate (no existing secrets)

This is *alpha* code, and you should not trust this with your data.



git-svn-id: http://secrets-for-android.googlecode.com/svn/branches/online-backup@94 1d24a8dc-1b16-11de-ac9f-1d9250355bca